--- apps/web/components/settings/general-settings/VAT-number-validation/VatNumberInfo.vue
@@ -1,5 +1,0 @@
-<template>
-  <UiFormLabel class="mb-6">
-    Do you want to check the validity of your customers’ VAT number? The VAT number is checked whenever an address is changed and upon completion of an order. The validation is carried out via the external service VIES.
-  </uiformlabel>
-</template>
